+++ lib/clean.rb
@@ -2,11 +2,9 @@
 
 require_relative "clean/version"
 require_relative 'clean/sentry'
-require_relative 'clean/validator'
 require_relative 'clean/service'
 require_relative 'clean/gateway'
-require_relative 'clean/service_chain'
-require_relative 'clean/service_adapter'
+require_relative 'clean/extra'
 
 module Clean