+++ lib/clean/extra.rb
@@ -1,0 +1,3 @@
+require_relative 'extra/validator'
+require_relative 'extra/adapter'
+require_relative 'extra/chain'