+++ test/clean/extra/test_chain.rb
@@ -1,11 +1,8 @@
-require_relative '../test_helper'
-require 'clean/service'
-require 'clean/service_chain'
-include Clean
+require_relative '../../test_helper'
 
-describe ServiceChain do
+describe Chain do
 
-  class ServiceChain
+  class Chain
     attr_reader :context
     attr_reader :presenter
     public_class_method :new
@@ -22,18 +19,18 @@ describe ServiceChain do
 
   describe '#new' do
     it 'must require *services' do
-      assert_raises(ArgumentError) { ServiceChain.() }
+      assert_raises(ArgumentError) { Chain.() }
     end
 
     it 'must allow empty **context' do
-      chain = ServiceChain.new(dummy, dummy)
+      chain = Chain.new(dummy, dummy)
       assert_equal Hash[], chain.context
       assert_equal Hash[], chain.presenter
     end
 
     let(:presenter) { {dummy => :d} }
     it 'must extract :presenter' do
-      service = ServiceChain.new(dummy, dummy, present: [dummy, :d])
+      service = Chain.new(dummy, dummy, present: [dummy, :d])
       assert_equal presenter, service.presenter
       refute service.context[:present]
     end
@@ -41,28 +38,28 @@ describe ServiceChain do
 
   describe 'presenter' do
     it 'must require even array' do
-      err1 = -> {ServiceChain.new(dummy, dummy, present: [dummy])}
-      err2 = -> {ServiceChain.new(dummy, dummy, present: [dummy, :d, dummy])}
+      err1 = -> {Chain.new(dummy, dummy, present: [dummy])}
+      err2 = -> {Chain.new(dummy, dummy, present: [dummy, :d, dummy])}
       assert_raises(ArgumentError) { err1.call }
       assert_raises(ArgumentError) { err2.call }
     end
 
     it 'must present result in :context' do
-      result = ServiceChain.(dummy, dummy, present: [dummy, :d])
+      result = Chain.(dummy, dummy, present: [dummy, :d])
       assert_includes result, :d
       assert_equal 42, result[:d]
     end
 
-    it 'must raise Failure unless result is a Hash or "presented"' do
-      err1 = -> { ServiceChain.(dummy, dummy) }
-      assert_raises(Service::Failure) { err1.call }
+    it 'must fail unless result is a Hash or "presented"' do
+      err1 = -> { Chain.(dummy, dummy) }
+      assert_raises(Chain::Failure) { err1.call }
 
       regular = Class.new(Service) do
         def call
           {a: 42}
         end
       end
-      ServiceChain.(regular, regular)
+      Chain.(regular, regular)
     end
   end
 
@@ -76,11 +73,11 @@ describe ServiceChain do
 
     let(:kwargs) { { a: 1, b: 2 } }
     it 'must require input from :context' do
-      chain = ServiceChain.new(serv1, serv1, **kwargs)
+      chain = Chain.new(serv1, serv1, **kwargs)
       assert_equal kwargs, chain.kwargs(serv1)
 
-      chain = ServiceChain.new(serv1, serv1, a: 1)
-      assert_raises(Service::Failure) { chain.kwargs(serv1) }
+      chain = Chain.new(serv1, serv1, a: 1)
+      assert_raises(Chain::Failure) { chain.kwargs(serv1) }
     end
 
     # @todo it 'maybe it must search deep?'
@@ -104,7 +101,7 @@ describe ServiceChain do
     }
 
     it 'must chain service1 and service2' do
-      assert_output("create\nupdate\n") { ServiceChain.(service1, service2) }
+      assert_output("create\nupdate\n") { Chain.(service1, service2) }
     end
   end
 
@@ -139,7 +136,7 @@ describe ServiceChain do
     }
 
     it 'must return ?' do
-      result = ServiceChain.(
+      result = Chain.(
         s1, s3, s2,
         a: 1, b: 2,
         present: [s1, :c, s2, :d]