+++ test/test_helper.rb
@@ -2,5 +2,6 @@
 
 $LOAD_PATH.unshift File.expand_path("../lib", __dir__)
 require "clean"
+include Clean
 
 require "minitest/autorun"